requirements.txt
Flask==2.3.2
gunicorn==20.0.4
numpy==1.23.5
nltk==3.8.1
scikit-learn==1.2.1
tensorflow==2.12.0
torch==2.0.1
Pillow==9.2.0
keras==2.12.0